Modularize hero creation flow
